+++ src/afml_system/utils/__init__.py
@@ -1,0 +1,27 @@
+"""
+PRADO9_EVO Utilities Module
+
+Common utilities for the PRADO trading system.
+"""
+
+from .paths import (
+    get_prado_root,
+    get_config_dir,
+    get_evo_dir,
+    get_models_dir,
+    get_live_dir,
+    get_portfolio_dir,
+    get_logs_dir,
+    migrate_from_home_dir
+)
+
+__all__ = [
+    'get_prado_root',
+    'get_config_dir',
+    'get_evo_dir',
+    'get_models_dir',
+    'get_live_dir',
+    'get_portfolio_dir',
+    'get_logs_dir',
+    'migrate_from_home_dir'
+]